requirements.txt
fastapi~=0.75.2
python-dotenv~=0.20.0
pymongo~=4.1.1